examples/a2a-demo/pyproject.toml
[project]
name = "a2a-demo"
version = "0.1.0"
description = "A2A cross-framework demo — standalone Legal Advisor agent"
requires-python = ">=3.13"
dependencies = [
    "agent-framework>=1.0.0rc5",
    "agent-framework-a2a>=1.0.0b260319",
    "a2a-sdk[http-server]",
    "azure-identity>=1.15.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "uvicorn>=0.30.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
